requirements.txt
loguru==0.5.3
pyTelegramBotAPI==4.1.1
python-dotenv==0.19.2
python-telegram-bot-calendar==1.0.5
requests==2.26.0
peewee==3.14.8